Removing api package, moving into network
