+++ src/main/kotlin/com/safehill/kclient/models/serde/ServerAssetSerializer.kt
@@ -1,4 +1,4 @@
-package com.safehill.kclient.api.serde
+package com.safehill.kclient.models.serde
 
 import com.safehill.kclient.models.dtos.AssetOutputDTO
 import com.safehill.kclient.models.dtos.AssetVersionOutputDTO
