+++ src/main/kotlin/com/safehill/kclient/network/LocalServerInterface.kt
@@ -1,6 +1,5 @@
 package com.safehill.kclient.network
 
-import com.safehill.kclient.api.SafehillApi
 import com.safehill.kclient.models.dtos.MessageOutputDTO
 import com.safehill.kclient.models.users.RemoteUser
 import com.safehill.kclient.models.dtos.ConversationThreadOutputDTO
