+++ src/main/kotlin/com/safehill/kclient/network/ServerProxyInterface.kt
@@ -1,6 +1,5 @@
 package com.safehill.kclient.network
 
-import com.safehill.kclient.api.SafehillApi
 import com.safehill.kclient.models.users.ServerUser
 
 interface ServerProxyInterface : SafehillApi {
