+++ src/test/kotlin/com/safehill/kclient/api/serde/ISO8601DateSerializerTests.kt
@@ -1,5 +1,7 @@
 package com.safehill.kclient.api.serde
 
+import com.safehill.kclient.models.serde.toIso8601Date
+import com.safehill.kclient.models.serde.toIso8601String
 import java.util.*
 import kotlin.test.Test
 import kotlin.test.assertEquals
